# api/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_erhan_falcon_api_internal_models.Program:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      rules:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_erhan_falcon_api_internal_models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  internal_handlers.errorResp:
    properties:
      error:
        type: string
    type: object
  internal_handlers.loginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_handlers.loginResp:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        properties:
          email:
            type: string
          id:
            type: integer
        type: object
    type: object
...
info:
  contact: {}
  description: Bug bounty & continuous recon backend.
  title: Falcon API
  version: 0.1.0
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.loginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      summary: Sign in with email + password
      tags:
      - auth
  /api/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/import/bugcrowd:
    post:
      consumes:
      - text/plain
      parameters:
      - description: raw HTML
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.Program'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Import Bugcrowd program from HTML
      tags:
      - programs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token issued by /api/auth/login. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"